Atiende review: cambio int por float

+++ operaciones.c
@@ -2,7 +2,7 @@
 
 /* ── sumar — ya implementada, referencia de estilo ───────────────────────── */
 
-int sumar(const int a, const int b) {
+int sumar(const float a, const float b) {
     return a + b;
 }
